refactor: renamed MenubarMenu to MenubarSubmenu for clarity
